Fix : JourneyService 수정 updateJourney log 수정, deleteJourney 호출 시 장소 삭제 추가

+++ server/src/main/java/onde/there/journey/service/JourneyService.java
@@ -16,18 +16,22 @@ import lombok.extern.slf4j.Slf4j;
 import onde.there.domain.Journey;
 import onde.there.domain.JourneyTheme;
 import onde.there.domain.Member;
+import onde.there.domain.Place;
 import onde.there.domain.type.JourneyThemeType;
 import onde.there.dto.journy.JourneyDto;
 import onde.there.dto.journy.JourneyDto.DetailResponse;
 import onde.there.dto.journy.JourneyDto.JourneyListResponse;
 import onde.there.dto.journy.JourneyDto.UpdateRequest;
 import onde.there.dto.journy.JourneyDto.UpdateResponse;
+import onde.there.exception.PlaceException;
+import onde.there.exception.type.ErrorCode;
 import onde.there.image.service.AwsS3Service;
 import onde.there.journey.exception.JourneyException;
 import onde.there.journey.repository.JourneyRepository;
 import onde.there.journey.repository.JourneyRepositoryImpl;
 import onde.there.journey.repository.JourneyThemeRepository;
 import onde.there.member.repository.MemberRepository;
+import onde.there.place.repository.PlaceRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
@@ -42,6 +46,7 @@ public class JourneyService {
 	private final MemberRepository memberRepository;
 	private final JourneyRepositoryImpl journeyRepositoryImpl;
 	private final AwsS3Service awsS3Service;
+	private final PlaceRepository placeRepository;
 
 	@Transactional
 	public JourneyDto.CreateResponse createJourney(
@@ -192,6 +197,13 @@ public class JourneyService {
 		List<JourneyTheme> journeyThemeTypeList = journeyThemeRepository
 			.findAllByJourneyId(journey.getId());
 
+		List<Place> list = placeRepository.findAllByJourneyOrderByPlaceTimeAsc(journey);
+
+		if (list.size() == 0) {
+			throw new PlaceException(ErrorCode.DELETED_NOTING);
+		}
+
+		placeRepository.deleteAll(list);
 		awsS3Service.deleteFile(journey.getJourneyThumbnailUrl());
 		journeyThemeRepository.deleteAll(journeyThemeTypeList);
 		journeyRepository.delete(journey);
@@ -231,8 +243,8 @@ public class JourneyService {
 					JourneyThemeType.findByTheme(inputJourneyTheme))
 				.build();
 			journeyThemeRepository.save(journeyTheme);
-			log.info("updateJourney() : journeyTheme 수정 완료");
 		}
+			log.info("updateJourney() : journeyTheme 수정 완료");
 
 		journey.setTitle(request.getTitle());
 		journey.setStartDate(request.getStartDate());
